+++ modules/rocksdb-common/src/main/java/org/apache/ignite/internal/rocksdb/RocksBiPredicate.java
@@ -1,0 +1,36 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.ignite.internal.rocksdb;
+
+import org.rocksdb.RocksDBException;
+
+/**
+ * BiPredicate that can throw {@link RocksDBException}.
+ */
+@FunctionalInterface
+public interface RocksBiPredicate {
+    /**
+     * Evaluates the predicate on the given key and the given value.
+     *
+     * @param key   Key.
+     * @param value Value.
+     * @return {@code true} if the input argument matches the predicate, otherwise {@code false}.
+     * @throws RocksDBException If failed to test the key-value pair.
+     */
+    boolean test(byte[] key, byte[] value) throws RocksDBException;
+}
